+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project follows [Semantic Versioning](https://semver.org/).
 
+## 1.1.3
+
+### Changed
+
+- Set `ShakeContainer.config` default value to `ShakePreset.error`, so users can omit `config` when
+  using `ShakeContainer` or `ShakeText`.
+
+---
+
 ## 1.1.2
 
 ### Changed